% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate data}
\usage{
validate(data, schema)
}
\arguments{
\item{data}{file name}

\item{schema}{file name}
}
\value{
Boolean
}
\description{
Validate data
}
\examples{
validate("data.csv", "schema.yaml")
}
